Migrated permissions from settings fragment
