+++ app/src/main/java/org/broeuschmeul/android/gps/usb/provider/UsbGpsBaseActivity.java
@@ -1,15 +1,22 @@
 package org.broeuschmeul.android.gps.usb.provider;
 
+import android.Manifest;
 import android.app.AlertDialog;
 import android.app.NotificationManager;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.SharedPreferences;
+import android.content.pm.PackageManager;
+import android.os.Build;
 import android.os.Bundle;
+import android.preference.CheckBoxPreference;
 import android.preference.PreferenceFragment;
 import android.preference.PreferenceManager;
+import android.preference.SwitchPreference;
 import android.provider.Settings;
+import android.support.annotation.NonNull;
+import android.support.v4.content.ContextCompat;
 import android.support.v7.app.AppCompatActivity;
 import android.view.MenuItem;
 
@@ -24,7 +31,7 @@ import android.view.MenuItem;
  * the nested settings.
  */
 
-public abstract class UsbGpsActivity extends AppCompatActivity implements
+public abstract class UsbGpsBaseActivity extends AppCompatActivity implements
         USBGpsSettingsFragment.PreferenceScreenListener,
         SharedPreferences.OnSharedPreferenceChangeListener {
 
@@ -34,6 +41,10 @@ public abstract class UsbGpsActivity extends AppCompatActivity implements
     private boolean shouldInitialise = true;
 
     private int resSettingsHolder;
+    private boolean tryingToStart;
+
+    private static final int LOCATION_REQUEST = 238472383;
+    private static final int STORAGE_REQUEST = 8972842;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -42,6 +53,13 @@ public abstract class UsbGpsActivity extends AppCompatActivity implements
 
         if (savedInstanceState != null) {
             shouldInitialise = false;
+        }
+
+        if (!hasPermission(Manifest.permission.ACCESS_FINE_LOCATION)) {
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+                requestPermissions(new String[]{Manifest.permission.ACCESS_FINE_LOCATION},
+                        LOCATION_REQUEST);
+            }
         }
 
         super.onCreate(savedInstanceState);
@@ -110,14 +128,135 @@ public abstract class UsbGpsActivity extends AppCompatActivity implements
         }
     }
 
+    /**
+     * Checks if the applications has the given runtime permission
+     * @param perm
+     * @return
+     */
+    private boolean hasPermission(String perm) {
+        return (
+                PackageManager.PERMISSION_GRANTED ==
+                        ContextCompat.checkSelfPermission(this, perm)
+        );
+    }
 
+    /**
+     * Android 6.0 requires permissions to be accepted at runtime
+     */
+    @Override
+    public void onRequestPermissionsResult(int requestCode,
+                                           @NonNull String[] permissions,
+                                           @NonNull int[] grantResults) {
+        if (requestCode == LOCATION_REQUEST) {
+            if (hasPermission(permissions[0])) {
+                if (tryingToStart) {
+                    tryingToStart = false;
+
+                    Intent serviceIntent = new Intent(this, USBGpsProviderService.class);
+                    serviceIntent.setAction(USBGpsProviderService.ACTION_START_GPS_PROVIDER);
+                    startService(serviceIntent);
+                }
+
+            } else {
+                tryingToStart = false;
+                sharedPreferences.edit().putBoolean(USBGpsProviderService.PREF_START_GPS_PROVIDER, false)
+                        .apply();
+                new AlertDialog.Builder(this)
+                        .setMessage("Location access needs to be enabled for this app to function")
+                        .setPositiveButton(android.R.string.ok, null)
+                        .show();
+
+            }
+
+        } else if (requestCode == STORAGE_REQUEST) {
+            if (hasPermission(permissions[0])) {
+                Intent serviceIntent = new Intent(this, USBGpsProviderService.class);
+                serviceIntent.setAction(USBGpsProviderService.ACTION_START_TRACK_RECORDING);
+                startService(serviceIntent);
+
+            } else {
+                sharedPreferences
+                        .edit()
+                        .putBoolean(USBGpsProviderService.PREF_TRACK_RECORDING, false)
+                        .apply();
+
+                new AlertDialog.Builder(this).setMessage(
+                        "In order to write a track file, the app need storage permission")
+                        .setPositiveButton(android.R.string.ok, null)
+                        .show();
+
+            }
+        }
+    }
+
+
+    /**
+     * Handles service attributes changing and requesting permissions
+     */
     @Override
     public void onSharedPreferenceChanged(SharedPreferences sharedPreferences, String key) {
         switch (key) {
             case USBGpsProviderService.PREF_START_GPS_PROVIDER: {
+                boolean val = sharedPreferences.getBoolean(key, false);
+
                 if (!sharedPreferences
                         .getBoolean(USBGpsProviderService.PREF_START_GPS_PROVIDER, false)) {
                     showStopDialog();
+                }
+
+                if (val) {
+                    if (hasPermission(Manifest.permission.ACCESS_FINE_LOCATION)) {
+                        Intent serviceIntent = new Intent(this, USBGpsProviderService.class);
+                        serviceIntent.setAction(USBGpsProviderService.ACTION_START_GPS_PROVIDER);
+                        startService(serviceIntent);
+
+
+                    } else {
+                        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+                            tryingToStart = true;
+                            requestPermissions(new String[]{Manifest.permission.ACCESS_FINE_LOCATION},
+                                    LOCATION_REQUEST);
+                        }
+                    }
+
+                } else {
+                    Intent serviceIntent = new Intent(this, USBGpsProviderService.class);
+                    serviceIntent.setAction(USBGpsProviderService.ACTION_STOP_GPS_PROVIDER);
+                    startService(serviceIntent);
+                }
+
+                break;
+            }
+            case USBGpsProviderService.PREF_TRACK_RECORDING: {
+                boolean val = sharedPreferences.getBoolean(key, false);
+
+                if (val) {
+                    if (!hasPermission(Manifest.permission.WRITE_EXTERNAL_STORAGE)) {
+                        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+                            requestPermissions(new String[]{Manifest.permission.WRITE_EXTERNAL_STORAGE},
+                                    STORAGE_REQUEST);
+                        }
+                    } else {
+                        Intent serviceIntent = new Intent(this, USBGpsProviderService.class);
+                        serviceIntent.setAction(USBGpsProviderService.ACTION_START_TRACK_RECORDING);
+                        startService(serviceIntent);
+                    }
+
+                } else {
+                    Intent serviceIntent = new Intent(this, USBGpsProviderService.class);
+                    serviceIntent.setAction(USBGpsProviderService.ACTION_STOP_TRACK_RECORDING);
+                    startService(serviceIntent);
+                }
+
+                break;
+            }
+            case USBGpsProviderService.PREF_SIRF_GPS: {
+                if (sharedPreferences.getBoolean(USBGpsProviderService.PREF_START_GPS_PROVIDER, false)) {
+                    if (sharedPreferences.getBoolean(USBGpsProviderService.PREF_SIRF_GPS, false)) {
+                        Intent configIntent = new Intent(this, USBGpsProviderService.class);
+                        configIntent.setAction(USBGpsProviderService.ACTION_CONFIGURE_SIRF_GPS);
+                        startService(configIntent);
+                    }
                 }
             }
         }
